Initial login and registration pages.
Initial login and registration pages. Functional mockups, need to add email functionality as well as redirect on log in, etc. A cookie is made which can be used on other pages to reference user on login. Assumes use of local database for now, not using dbConnect.php. Need to be stylistically updated. Refinement to come.