Updates to the styling
This includes updates to the styling of the page which makes the pictures smaller, and moves the text to be aside the pictures instead of beneath, making it look more like a normal online store. I've also gotten images for the products in the store, although they can easily be switched to the simple Product picture blocks.